+++ cumulus_library/cli.py
@@ -10,14 +10,12 @@ from pathlib import Path, PosixPath
 from typing import Dict, List, Optional
 
 import pyathena
-import requests
 
 from pyathena.pandas.cursor import PandasCursor
-from rich.progress import Progress, TaskID
 
 from cumulus_library.cli_parser import get_parser
-from cumulus_library.helper import get_progress_bar
 from cumulus_library.study_parser import StudyManifestParser
+from cumulus_library.upload import upload_files
 
 
 # ** Don't delete! **
@@ -161,68 +159,6 @@ def create_template(path: str) -> None:
         source_path = Path(Path(__file__).resolve().parents[0], source)
         dest_path = Path(abs_path, dest)
         dest_path.write_bytes(source_path.read_bytes())
-
-
-def upload_data(
-    progress: Progress, file_upload_progress: TaskID, file_path: Path, args: dict
-):
-    """Fetches presigned url and uploads file to aggregator"""
-    study = file_path.parts[-2]
-    file_name = file_path.parts[-1]
-    progress.update(file_upload_progress, description=f"Uploading {study}/{file_name}")
-    subscription = file_name.split(".")[0]
-    prefetch_res = requests.post(
-        args["url"],
-        json={
-            "study": study,
-            "data_package": subscription,
-            "filename": f"{args['user']}_{file_name}",
-        },
-        auth=(args["user"], args["id"]),
-        timeout=60,
-    )
-    if args["preview"]:
-        print("prefetch request")
-        print("headers", prefetch_res.request.headers)
-        print("body", prefetch_res.request.body, "\n")
-        print("response")
-        print(prefetch_res.json(), "\n")
-
-    if prefetch_res.status_code != 200:
-        print("Invalid user/site id")
-        raise requests.RequestException(response=prefetch_res)
-    res_body = prefetch_res.json()
-
-    with open(file_path, "rb") as data_file:
-        files = {"file": (file_name, data_file)}
-        upload_req = requests.Request(
-            "POST", res_body["url"], data=res_body["fields"], files=files
-        ).prepare()
-        if not args["preview"]:
-            s = requests.Session()
-            upload_res = s.send(upload_req, timeout=60)
-            if upload_res.status_code != 204:
-                print(f"Error uploading {study}/{file_name}")
-                raise requests.RequestException(response=upload_res)
-        else:
-            print("upload_req")
-            print("headers", upload_req.headers)
-            print("body", upload_req.body, "\n")
-    progress.update(file_upload_progress, advance=1)
-
-
-def upload_files(args: dict):
-    """Wrapper to prep files & console output"""
-    if args["export_dir"] is None:
-        args["export_dir"] = Path(__file__).resolve().parent / "data_export"
-    file_paths = list(args["export_dir"].glob("**/*.parquet"))
-    num_uploads = len(file_paths)
-    if not args["user"] or not args["id"]:
-        sys.exit("user/id not provided, please pass --user and --id")
-    with get_progress_bar() as progress:
-        file_upload_progress = progress.add_task("Uploading", total=num_uploads)
-        for file_path in file_paths:
-            upload_data(progress, file_upload_progress, file_path, args)
 
 
 def get_study_dict(alt_dir_paths: List) -> Optional[Dict[str, PosixPath]]:
